requirements.txt

langchain-groq==0.3.8
langchain-core==0.3.77
langchain-community==0.3.30
pypdf==6.1.1
sentence-transformers==5.1.1
chromadb==1.1.0
Flask==3.1.2
python-dotenv